refactor(main): main.ts を index.ts / windows.ts / ipcHandlers.ts / services に分割
コードは移動のみで挙動は不変。窓に依存しないハンドラは ipcHandlers.ts、
窓生成と窓依存ハンドラは windows.ts、更新チェック・ダウンロード・
nicommons API は services/ へ。webpack の main エントリを index.ts に変更。
