+++ src/main/windows.ts
@@ -1,0 +1,231 @@
+import {
+  BrowserWindow,
+  app,
+  dialog,
+  ipcMain,
+  Menu,
+  nativeTheme,
+  shell,
+} from 'electron';
+import prompt from 'electron-prompt';
+import { createIPCHandler } from 'electron-trpc/main';
+import windowStateKeeper from 'electron-window-state';
+import path from 'node:path';
+import { IPC_CHANNELS } from '../common/ipc';
+import type Config from '../lib/Config';
+import { router } from './api';
+import { runAutoUpdate } from './services/appUpdate';
+import { downloadFile } from './services/download';
+
+declare const SPLASH_WINDOW_WEBPACK_ENTRY: string;
+declare const MAIN_WINDOW_WEBPACK_ENTRY: string;
+declare const MAIN_WINDOW_PRELOAD_WEBPACK_ENTRY: string;
+declare const ABOUT_WINDOW_WEBPACK_ENTRY: string;
+declare const ABOUT_WINDOW_PRELOAD_WEBPACK_ENTRY: string;
+
+const isDevEnv = process.env.NODE_ENV === 'development';
+
+const icon =
+  process.platform === 'linux'
+    ? path.join(__dirname, '../icon/apm1024.png')
+    : undefined;
+
+/**
+ * Launch the app.
+ * @param {Config} config - The config instance.
+ */
+export async function launch(config: Config) {
+  await runAutoUpdate(config, isDevEnv);
+
+  const splashWindow = new BrowserWindow({
+    width: 640,
+    height: 360,
+    center: true,
+    frame: false,
+    show: false,
+    icon: icon,
+  });
+
+  splashWindow.once('ready-to-show', () => {
+    splashWindow.show();
+  });
+
+  void splashWindow.loadURL(SPLASH_WINDOW_WEBPACK_ENTRY);
+
+  const mainWindowState = windowStateKeeper({
+    defaultWidth: 800,
+    defaultHeight: 600,
+  });
+
+  const getTitleBarColor = () => {
+    return {
+      color: nativeTheme.shouldUseDarkColors ? '#2b3035' : '#f8f9fa',
+      symbolColor: nativeTheme.shouldUseDarkColors ? '#dee2e6' : '#212529',
+    };
+  };
+
+  const mainWindow = new BrowserWindow({
+    x: mainWindowState.x,
+    y: mainWindowState.y,
+    width: mainWindowState.width,
+    height: mainWindowState.height,
+    minWidth: 320,
+    minHeight: 240,
+    show: false,
+    icon: icon,
+    titleBarStyle: 'hidden',
+    titleBarOverlay: getTitleBarColor(),
+    webPreferences: {
+      preload: MAIN_WINDOW_PRELOAD_WEBPACK_ENTRY,
+      sandbox: false,
+    },
+  });
+
+  Menu.setApplicationMenu(null);
+
+  mainWindow.webContents.on('will-navigate', async (event, url) => {
+    if (url.match(/^http/)) {
+      event.preventDefault();
+      await shell.openExternal(url);
+    }
+  });
+  mainWindow.webContents.setWindowOpenHandler((details) => {
+    if (details.url.match(/^http/)) {
+      void shell.openExternal(details.url);
+    }
+    return { action: 'deny' };
+  });
+
+  nativeTheme.on('updated', () => {
+    mainWindow.setTitleBarOverlay(getTitleBarColor());
+  });
+
+  mainWindow.once('show', () => {
+    mainWindowState.manage(mainWindow);
+  });
+
+  ipcMain.handle(IPC_CHANNELS.OPEN_ABOUT_WINDOW, async () => {
+    const aboutPath = ABOUT_WINDOW_WEBPACK_ENTRY;
+    const aboutWindow = new BrowserWindow({
+      width: 480,
+      height: 360,
+      frame: false,
+      resizable: false,
+      modal: true,
+      parent: mainWindow,
+      icon: icon,
+      webPreferences: {
+        preload: ABOUT_WINDOW_PRELOAD_WEBPACK_ENTRY,
+        sandbox: false,
+      },
+    });
+    createIPCHandler({ router, windows: [aboutWindow] });
+    aboutWindow.once('close', () => {
+      if (!aboutWindow.isDestroyed()) {
+        aboutWindow.destroy();
+      }
+    });
+    aboutWindow.once('ready-to-show', () => {
+      aboutWindow.show();
+    });
+    void aboutWindow.loadURL(aboutPath);
+  });
+
+  ipcMain.handle(IPC_CHANNELS.MIGRATION1TO2_CONFIRM_DIALOG, async () => {
+    return (
+      await dialog.showMessageBox(mainWindow, {
+        title: '確認',
+        message: `お使いのバージョンのapmは現在設定されているデータ取得先に対応しておりません。新しいデータ取得先への移行が必要です。`,
+        type: 'warning',
+        buttons: [
+          'キャンセル',
+          '新しいデータ取得先を入力する',
+          'デフォルトのデータ取得先を使う',
+        ],
+        cancelId: 0,
+      })
+    ).response;
+  });
+
+  ipcMain.handle(IPC_CHANNELS.MIGRATION1TO2_DATAURL_INPUT_DIALOG, async () => {
+    return await prompt(
+      {
+        title: '新しいデータ取得先の入力',
+        label: '新しいデータ取得先のURL（例: https://example.com/data/）',
+        width: 500,
+        height: 300,
+        type: 'input',
+      },
+      mainWindow,
+    );
+  });
+
+  ipcMain.handle(IPC_CHANNELS.CHANGE_MAIN_ZOOM_FACTOR, (event, zoomFactor) => {
+    mainWindow.webContents.setZoomFactor(zoomFactor);
+  });
+
+  ipcMain.handle(IPC_CHANNELS.DOWNLOAD, async (event, url, options) => {
+    return await downloadFile(mainWindow, url, options);
+  });
+
+  ipcMain.handle(IPC_CHANNELS.OPEN_BROWSER, async (event, url, type) => {
+    const browserWindow = new BrowserWindow({
+      width: 800,
+      height: 600,
+      minWidth: 240,
+      minHeight: 320,
+      webPreferences: { sandbox: true },
+      parent: mainWindow,
+      modal: true,
+      icon: icon,
+    });
+
+    mainWindow.once('closed', () => {
+      if (!browserWindow.isDestroyed()) {
+        browserWindow.destroy();
+      }
+    });
+
+    void browserWindow.loadURL(url);
+
+    return await new Promise((resolve) => {
+      const history: string[] = [];
+
+      browserWindow.webContents.on('did-navigate', (e, url) => {
+        history.push(url);
+      });
+
+      browserWindow.webContents.session.once('will-download', (event, item) => {
+        if (!browserWindow.isDestroyed()) browserWindow.hide();
+
+        const ext = path.extname(item.getFilename());
+        const dir = path.join(app.getPath('userData'), 'Data');
+        if (['.zip', '.lzh', '.7z', '.rar'].includes(ext)) {
+          item.setSavePath(
+            path.join(dir, type, 'archive/', item.getFilename()),
+          );
+        } else {
+          item.setSavePath(path.join(dir, type, item.getFilename()));
+        }
+
+        item.once('done', () => {
+          history.push(...item.getURLChain(), item.getFilename());
+          resolve({ savePath: item.getSavePath(), history: history });
+          if (!browserWindow.isDestroyed()) browserWindow.close();
+        });
+      });
+
+      browserWindow.once('closed', () => {
+        resolve(null);
+      });
+    });
+  });
+
+  setTimeout(() => {
+    mainWindow.show();
+    splashWindow.hide();
+    splashWindow.destroy();
+  }, 2000);
+
+  void mainWindow.loadURL(MAIN_WINDOW_WEBPACK_ENTRY);
+}
